Correct search deactivation bug
Correct an issue on SearchTrack component.
When a track search was done, the result was displayed in a dropdown list below the search input. This dropdown list was shown when the search input was focused and unmounted when the input was not focused anymore. However, clicking on the result list (to add the track to the playlist, for example) would unfocus the search input and close the dropdown list.
This correction aims to resolve that issue and keep the dropdown list open when clicking on it.

1. Create a custom hook to execute a function when clicking outside a given element.
2. Use this custom hook to deactivate the dropdown list when clicking outside the search element (the search element include the dropdown list) instead of using onblur on the sole input to deactivate the list.

Also correct the desactivateSearch function to deactivateSearch.