Keep the current playlist selected when a track is add or removed
When a track was added or removed to/from a playlist, the playlist was refreshed to update the list of playlists and tracks. When refreshed, the index of the selected playlist was re-initialized to 0. So when you added a song to the fourth playlist, the playlist selected changed to become the first one.
Split the playlist refresh and the playlist re-initialization.
* Refresh when a song is added/removed to/from a playlist or when a new playlist is created.
* Re-initialize index only when a playlist is created.